slider webp

+++ app/Http/Controllers/SliderController.php
@@ -73,33 +73,104 @@ class SliderController extends Controller
     {
         // Validasi request
         $validator = Validator::make($request->all(), $this->validator->getRules(), $this->validator->getMessageBag()->toArray());
-
+    
         if ($validator->fails()) {
             return response()->json(['errors' => $validator->errors()], 422);
         }
-
+    
         $input = $request->all();  // Pindahkan ini ke bawah validasi
-
+    
         if ($request->hasFile('gambar')) {
             $image = $request->file('gambar');
             $destinationPath = 'upload/slider/';
-            $imageName = time() . '_' . $image->getClientOriginalName();
-            $image->move($destinationPath, $imageName);
-            $input['gambar'] = $imageName;
+            
+            // Konversi gambar ke WebP
+            $imageName = $this->convertImageToWebp($image, $destinationPath);
+            if ($imageName) {
+                $input['gambar'] = $imageName;
+            } else {
+                return response()->json(['error' => 'Gagal mengonversi gambar ke WebP'], 500);
+            }
         }
-
+    
         // Simpan data spp ke database menggunakan fill()
         $slider = new Slider();
         $slider->fill($input);
         $slider->save();
-
+    
         $loggedInUserId = Auth::id();
-
+    
         // Simpan log histori untuk operasi Create dengan user_id yang sedang login
         $this->simpanLogHistori('Create', 'Slider', $slider->id, $loggedInUserId, null, json_encode($slider));
-
+    
         return response()->json(['message' => 'Data Berhasil Disimpan']);
     }
+    
+    // Fungsi untuk mengonversi gambar ke WebP
+    private function convertImageToWebp($image, $destinationPath)
+    {
+        // Pastikan direktori tujuan ada
+        if (!file_exists(public_path($destinationPath))) {
+            mkdir(public_path($destinationPath), 0777, true);
+        }
+    
+        // Ambil nama file asli dan ekstensinya
+        $originalFileName = $image->getClientOriginalName();
+    
+        // Ambil tipe MIME dari gambar
+        $imageMimeType = $image->getMimeType();
+    
+        // Filter hanya tipe MIME gambar yang didukung (misalnya, image/jpeg, image/png, dll.)
+        if (strpos($imageMimeType, 'image/') === 0) {
+            // Gabungkan waktu dengan nama file asli
+            $imageName = date('YmdHis') . '_' . str_replace(' ', '_', $originalFileName);
+    
+            // Simpan gambar asli ke tujuan yang diinginkan
+            $image->move(public_path($destinationPath), $imageName);
+    
+            // Path gambar asli
+            $sourceImagePath = public_path($destinationPath . $imageName);
+    
+            // Path untuk menyimpan gambar WebP
+            $webpImagePath = $destinationPath . pathinfo($imageName, PATHINFO_FILENAME) . '.webp';
+    
+            // Baca gambar asli dan konversi ke WebP jika tipe MIME didukung
+            switch ($imageMimeType) {
+                case 'image/jpeg':
+                    $sourceImage = @imagecreatefromjpeg($sourceImagePath);
+                    break;
+                case 'image/png':
+                    $sourceImage = @imagecreatefrompng($sourceImagePath);
+                    break;
+                    // Tambahkan jenis MIME lain jika diperlukan
+                default:
+                    // Jenis MIME tidak didukung, tangani kasus ini sesuai kebutuhan Anda
+                    return null;
+            }
+    
+            // Jika gambar asli berhasil dibaca
+            if ($sourceImage !== false) {
+                // Buat gambar baru dalam format WebP
+                imagewebp($sourceImage, public_path($webpImagePath));
+    
+                // Hapus gambar asli dari memori
+                imagedestroy($sourceImage);
+    
+                // Hapus file asli setelah konversi selesai
+                @unlink($sourceImagePath);
+    
+                // Kembalikan hanya nama file gambar WebP
+                return pathinfo($imageName, PATHINFO_FILENAME) . '.webp';
+            } else {
+                // Gagal membaca gambar asli, tangani kasus ini sesuai kebutuhan Anda
+                return null;
+            }
+        } else {
+            // Tipe MIME gambar tidak didukung, tangani kasus ini sesuai kebutuhan Anda
+            return null;
+        }
+    }
+    
 
 
     /**
@@ -135,40 +206,109 @@ class SliderController extends Controller
     public function update(Request $request, $id)
     {
         $validator = Validator::make($request->all(), $this->validator->getRules(), $this->validator->getMessageBag()->toArray());
-
+    
         if ($validator->fails()) {
             return response()->json(['errors' => $validator->errors()], 422);
         }
-
+    
         $slider = Slider::findOrFail($id);
-
+        $oldData = $slider->getOriginal();
+    
         $input = $request->except(['_token', '_method']); // Exclude unnecessary fields
-
+    
         if ($request->hasFile('gambar')) {
             $image = $request->file('gambar');
             $destinationPath = 'upload/slider/';
-            $imageName = time() . '_' . $image->getClientOriginalName();
-            $image->move($destinationPath, $imageName);
     
-            // Hapus file gambar lama jika ada
-            if ($slider->gambar && file_exists(public_path('upload/slider/' . $slider->gambar))) {
-                unlink(public_path('upload/slider/' . $slider->gambar));
+            // Konversi gambar ke WebP
+            $imageName = $this->convertImageToWebpUpdate($image, $destinationPath);
+            if ($imageName) {
+                // Hapus file gambar lama jika ada
+                if ($slider->gambar && file_exists(public_path('upload/slider/' . $slider->gambar))) {
+                    unlink(public_path('upload/slider/' . $slider->gambar));
+                }
+                $input['gambar'] = $imageName;
+            } else {
+                return response()->json(['error' => 'Gagal mengonversi gambar ke WebP'], 500);
             }
-    
-            $input['gambar'] = $imageName;
         }
     
         // Update slider data di database
         $slider->update($input);
-
+    
         $loggedInUserId = Auth::id();
-
+    
         // Simpan log histori untuk operasi Update dengan user_id yang sedang login
-        $this->simpanLogHistori('Update', 'Slider', $slider->id, $loggedInUserId, json_encode($slider->getOriginal()), json_encode($slider));
-
+        $this->simpanLogHistori('Update', 'Slider', $slider->id, $loggedInUserId, json_encode($oldData), json_encode($slider));
+    
         return response()->json(['message' => 'Data Berhasil Diupdate']);
     }
-
+    
+    // Fungsi untuk mengonversi gambar ke WebP
+    private function convertImageToWebpUpdate($image, $destinationPath)
+    {
+        // Pastikan direktori tujuan ada
+        if (!file_exists(public_path($destinationPath))) {
+            mkdir(public_path($destinationPath), 0777, true);
+        }
+    
+        // Ambil nama file asli dan ekstensinya
+        $originalFileName = $image->getClientOriginalName();
+    
+        // Ambil tipe MIME dari gambar
+        $imageMimeType = $image->getMimeType();
+    
+        // Filter hanya tipe MIME gambar yang didukung (misalnya, image/jpeg, image/png, dll.)
+        if (strpos($imageMimeType, 'image/') === 0) {
+            // Gabungkan waktu dengan nama file asli
+            $imageName = date('YmdHis') . '_' . str_replace(' ', '_', $originalFileName);
+    
+            // Simpan gambar asli ke tujuan yang diinginkan
+            $image->move(public_path($destinationPath), $imageName);
+    
+            // Path gambar asli
+            $sourceImagePath = public_path($destinationPath . $imageName);
+    
+            // Path untuk menyimpan gambar WebP
+            $webpImagePath = $destinationPath . pathinfo($imageName, PATHINFO_FILENAME) . '.webp';
+    
+            // Baca gambar asli dan konversi ke WebP jika tipe MIME didukung
+            switch ($imageMimeType) {
+                case 'image/jpeg':
+                    $sourceImage = @imagecreatefromjpeg($sourceImagePath);
+                    break;
+                case 'image/png':
+                    $sourceImage = @imagecreatefrompng($sourceImagePath);
+                    break;
+                    // Tambahkan jenis MIME lain jika diperlukan
+                default:
+                    // Jenis MIME tidak didukung, tangani kasus ini sesuai kebutuhan Anda
+                    return null;
+            }
+    
+            // Jika gambar asli berhasil dibaca
+            if ($sourceImage !== false) {
+                // Buat gambar baru dalam format WebP
+                imagewebp($sourceImage, public_path($webpImagePath));
+    
+                // Hapus gambar asli dari memori
+                imagedestroy($sourceImage);
+    
+                // Hapus file asli setelah konversi selesai
+                @unlink($sourceImagePath);
+    
+                // Kembalikan hanya nama file gambar WebP
+                return pathinfo($imageName, PATHINFO_FILENAME) . '.webp';
+            } else {
+                // Gagal membaca gambar asli, tangani kasus ini sesuai kebutuhan Anda
+                return null;
+            }
+        } else {
+            // Tipe MIME gambar tidak didukung, tangani kasus ini sesuai kebutuhan Anda
+            return null;
+        }
+    }
+
